requirements.txt
Django==1.11.28
ansible==2.6.18
appnope==0.1.0
asn1crypto==0.23.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.4
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
cryptography==2.8
decorator==4.1.2
diff-match-patch==20121119
django-crispy-forms==1.7.0
django-crontab==0.7.1
django-formtools==2.1
django-import-export==0.5.1
enum34==1.1.10
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
idna==2.6
ipaddress==1.0.18
ipython==5.5.0
ipython-genutils==0.2.0
jdcal==1.3
jieba==0.39
Jinja2==2.10.1
ldap3==2.4
MarkupSafe==1.0
odfpy==1.3.5
openpyxl==2.4.9
paramiko==2.4.2
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
ply==3.10
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.4.6
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.6.6
Pygments==2.2.0
pymongo==3.5.1
PyMySQL==0.9.3
PyNaCl==1.2.0
pysmi==0.1.3
pysnmp==4.3.9
python-nmap==0.6.1
pytz==2017.3
pyvmomi==6.5.0.2017.5.post1
PyYAML==5.3.1
redis==2.10.6
requests==2.20.0
scandir==1.6
simplegeneric==0.8.1
simplejson==3.13.2
six==1.12
tablib==0.12.1
traitlets==4.3.2
unicodecsv==0.14.1
urllib3==1.24.2
wcwidth==0.1.7
Whoosh==2.7.4
xadmin==0.6.1
xlrd==1.1.0
xlwt==1.3.0